Refactored loadDefaults function
Eliminated function calls for each individual material property and instead wrapped everything in a zip. Overall, trimmed the function by ~20 lines, but improved readability. Note - currently, the defaults are hard-coded in this function. It may be worthwhile to later pull them out to by read in from a separate text file; will make an issue.